Move onboarding to a dedicated module
