+++ settings.gradle
@@ -80,6 +80,7 @@ include ':simplified-ui-errorpage'
 include ':simplified-ui-images'
 include ':simplified-ui-navigation-api'
 include ':simplified-ui-navigation-tabs'
+include ':simplified-ui-onboarding'
 include ':simplified-ui-profiles'
 include ':simplified-ui-screen'
 include ':simplified-ui-settings'
