+++ simplified-app-vanilla/build.gradle
@@ -33,4 +33,6 @@ dependencies {
   annotationProcessor libraries.google_autovalue_processor
 
   vanillaWithProfilesImplementation project(":simplified-accounts-source-filebased")
+
+  implementation project(":simplified-migration-fake")
 }
