+++ simplified-ui-onboarding/gradle.properties
@@ -1,0 +1,4 @@
+POM_ARTIFACT_ID=org.librarysimplified.ui.onboarding
+POM_DESCRIPTION=Library Simplified (Onboarding)
+POM_NAME=org.librarysimplified.ui.onboarding
+POM_PACKAGING=aar
