+++ simplified-ui-onboarding/src/main/java/org/nypl/simplified/ui/onboarding/OnboardingStartScreenViewModel.kt
@@ -1,4 +1,4 @@
-package org.nypl.simplified.ui.splash
+package org.nypl.simplified.ui.onboarding
 
 import androidx.lifecycle.ViewModel
 import org.librarysimplified.services.api.Services
