+++ simplified-ui-onboarding/src/main/res/layout/onboarding_fragment.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<androidx.fragment.app.FragmentContainerView
+  xmlns:android="http://schemas.android.com/apk/res/android"
+  android:name="org.nypl.simplified.ui.onboarding.OnboardingStartScreenFragment"
+  android:id="@+id/onboarding_fragment_container"
+  android:layout_width="match_parent"
+  android:layout_height="match_parent">
+</androidx.fragment.app.FragmentContainerView>
