+++ simplified-ui-onboarding/src/main/res/layout/onboarding_start_screen.xml
@@ -1,0 +1,50 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+  android:layout_width="match_parent"
+  android:layout_height="match_parent"
+  android:gravity="center"
+  android:orientation="vertical">
+
+  <ImageView
+    android:id="@+id/selectionImage"
+    android:layout_width="272dp"
+    android:layout_height="wrap_content"
+    android:layout_marginBottom="32dp"
+    android:adjustViewBounds="true"
+    android:contentDescription="@null"
+    android:scaleType="centerInside" />
+
+  <Button
+    android:id="@+id/selectionButton"
+    android:layout_width="256dp"
+    android:layout_height="wrap_content"
+    android:layout_marginTop="16dp"
+    android:text="@string/selectionButton"
+    android:textSize="18sp"
+    android:textStyle="bold" />
+
+  <TextView
+    android:id="@+id/selectionTitle"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:layout_marginTop="4dp"
+    android:text="@string/selectionTitle"
+    android:textAppearance="@android:style/TextAppearance.Small" />
+
+  <Button
+    android:id="@+id/selectionAlternateButton"
+    android:layout_width="256dp"
+    android:layout_height="wrap_content"
+    android:layout_marginTop="16dp"
+    android:text="@string/selectionAlternateButton"
+    android:textSize="18sp"
+    android:textStyle="bold" />
+
+  <TextView
+    android:id="@+id/selectionAlternateTitle"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:layout_marginTop="4dp"
+    android:text="@string/selectionAlternateTitle"
+    android:textAppearance="@android:style/TextAppearance.Small" />
+</LinearLayout>
