+++ simplified-ui-onboarding/src/main/res/values/public.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<resources>
+  <public />
+</resources>
