+++ simplified-ui-onboarding/src/main/res/values/strings.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+
+  <!-- Onboarding start screen strings -->
+  <string name="selectionTitle">Read E-Books from Your Library</string>
+  <string name="selectionButton">Find Your Library</string>
+  <string name="selectionAlternateTitle">Browse the SimplyE Collection</string>
+  <string name="selectionAlternateButton">Add a Library Later</string>
+</resources>
